# newsArticle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PulseGurgaon News Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TimelineStep } from "./timelineStep";
import type { VocabWord } from "./vocabWord";

export interface NewsArticle {
  id: number;
  title: string;
  description: string;
  fullArticle: string;
  summaryPoints: string[];
  timeline: TimelineStep[];
  vocabulary: VocabWord[];
  category: string;
  image?: string | null;
  publishedAt: string;
  sourceUrl?: string | null;
  sourceName?: string | null;
  aiEnhanced: boolean;
  createdAt: string;
}