# newsListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PulseGurgaon News Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { NewsArticle } from "./newsArticle";

export interface NewsListResponse {
  articles: NewsArticle[];
  total: number;
  page: number;
  limit: number;
}