# openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PulseGurgaon News Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news articles
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: News articles list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsListResponse"
...
  /news/{id}:
    get:
      operationId: getNewsById
      tags: [news]
      summary: Get a single news article by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: News article detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsArticle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    NewsArticle:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        fullArticle:
          type: string
        summaryPoints:
          type: array
          items:
            type: string
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelineStep"
        vocabulary:
          type: array
          items:
            $ref: "#/components/schemas/VocabWord"
        category:
          type: string
        image:
          type: string
          nullable: true
        publishedAt:
          type: string
        sourceUrl:
          type: string
          nullable: true
        sourceName:
          type: string
          nullable: true
        aiEnhanced:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - title
        - description
        - fullArticle
        - summaryPoints
        - timeline
        - vocabulary
        - category
        - publishedAt
        - aiEnhanced
        - createdAt
    TimelineStep:
      type: object
      properties:
        step:
          type: number
        title:
          type: string
        description:
          type: string
      required:
        - step
        - title
        - description
    VocabWord:
      type: object
      properties:
        word:
          type: string
        meaning:
          type: string
      required:
        - word
        - meaning
    NewsListResponse:
      type: object
      properties:
        articles:
          type: array
          items:
            $ref: "#/components/schemas/NewsArticle"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - articles
        - total
        - page
        - limit
...